backend_python/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic==2.6.0
pydantic-settings==2.1.0
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-multipart==0.0.6
boto3==1.34.14
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyopenssl==24.0.0
cryptography==42.0.0
requests==2.31.0
numpy<2
rembg==2.0.50
pillow==10.2.0
onnxruntime==1.17.0


pandas==2.2.0
validate-docbr==1.10.0

# Logging and Metrics
python-json-logger==2.0.7
prometheus-fastapi-instrumentator==6.1.0

